+++ packages/ui/src/elements/PublishButton/ScheduleDrawer/buildUpcomingColumns.tsx
@@ -5,7 +5,7 @@ import React from 'react'
 
 import type { UpcomingEvent } from './types.js'
 
-import { formatDate } from '../../../utilities/formatDate.js'
+import { formatDate } from '../../../utilities/formatDocTitle/formatDateTitle.js'
 import { Button } from '../../Button/index.js'
 import { Pill } from '../../Pill/index.js'
