+++ packages/ui/src/utilities/formatDocTitle/formatDateTitle.ts
